requirements.txt
altair
pandas
streamlit
gsheetsdb
psycopg2-binary
